// SofTec_TMS/NWC.Web/src/app/TMS-Module/Models/contract-terms-violations-invoice.model.ts
export class ContractTermsViolationsInVoiceDTO {
    Id: number;
    ViolationNo: string;
    InvoiceNo: string;
    Value: number;
    VAT: number;
    ValueWithVAT: number;
    CreatedBy: string;
    CreatedDate: Date;
    ViolationId: number;
    ContractTermName: string;
    StationID: string;
    StationName: string;
    VehicleId: string;
    VehicleCode: string;
    VehiclePlateNo: string;
    DriverId: string;
    DriverName: string;
    DriverMobileNumber: string;
    DriverCode: string;
    CreatedByName: string;

}